requirements.txt
aiohttp==3.5.4
aiologger==0.3.0
aioredis==1.2.0
aio-pika==5.5.1
aiosmtplib==1.0.4
asyncpg==0.18.3
async-timeout==3.0.1
ujson==1.35
amqp==2.4.2
attrs==19.1.0
backcall==0.1.0
billiard==3.5.0.5
celery==4.2.1
chardet==3.0.4
decorator==4.3.2
Django==2.1.7
djangorestframework==3.9.2
hiredis==1.0.0
idna==2.8
idna-ssl==1.1.0
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
kombu==4.4.0
multidict==4.5.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
six==1.12.0
traitlets==4.3.2
typing-extensions==3.7.2
vine==1.2.0
wcwidth==0.1.7
yarl==1.3.0
websockets==6.0
prometheus-client==0.6.0
redis==3.2.1
requests==2.21.0
line-profiler==2.1.2
supervisor==4.0.2
raven==6.10.0
aiomysql==0.0.20
